requirements.txt
requests==2.32.3
cryptography==43.0.3
python-dotenv==1.0.1
openpyxl==3.1.5
python-dateutil==2.9.0
